Refactorization of pes_func, lsi management; tweaks and fixes

+++ lib/parallellinesearch.py
@@ -1,20 +1,21 @@
 #!/usr/bin/env python3
 
 from numpy import ndarray, array
-from os import makedirs
+from os import makedirs, path
 from dill import dumps
 from textwrap import indent
 
-from lib.util import get_fraction_error
+from lib.util import get_fraction_error, directorize
 from lib.parameters import ParameterSet
 from lib.hessian import ParameterHessian
 from lib.linesearch import LineSearch
+from lib.pessampler import PesSampler
 
 
 # Class for a bundle of parallel line-searches
-class ParallelLineSearch():
+class ParallelLineSearch(PesSampler):
 
-    ls_list = None  # list of line-search objects
+    ls_list = []  # list of line-search objects
     hessian = None  # hessian object
     Lambdas = None
     directions = None
@@ -26,23 +27,17 @@ class ParallelLineSearch():
     fraction = None
     structure = None  # eqm structure
     structure_next = None  # next structure
-    job_func = None
-    job_args = {}
-    analyze_func = None
     x_unit = None
     E_unit = None
-    # flags
     noisy = False  # flag whether deterministic or noisy
-    limits_set = False  # flag whether sensible line-search limits (windows, noises) are set
-    protected  = False  # whether the structure is considered frozen and protected from changes
-    generated = False  # whether structures have been generated
-    loaded = False  # has loaded line-search data
-    calculated = False  # new positions calculated
+    msg_setup = 'Setup: first set_structure() and set_hessian() with valid input'
+    msg_shiftep = 'Shifted: first set_windows() to define displacements'
+    msg_loaded = 'Loaded: first load_results() with valid input'
 
     def __init__(
         self,
-        hessian,
-        structure,
+        hessian = None,
+        structure = None,
         windows = None,
         window_frac = 0.25,
         noises = None,
@@ -52,36 +47,74 @@ class ParallelLineSearch():
         x_unit = 'A',
         E_unit = 'Ry',
         fraction = 0.025,
-        job_func = None,
-        job_args = {},
-        analyze_func = None,
-        mode = 'jobs',
-        pes_func = None,
+        mode = 'pes',  # (nexus|files|pes)
         **kwargs,
     ):
+        PesSampler.__init__(self, mode, **kwargs)
         self.x_unit = x_unit
         self.E_unit = E_unit
+        self.path = directorize(path)
         self.fraction = fraction
-        self.path = path
-        self.job_func = job_func
-        self.analyze_func = analyze_func
-        self.job_args = job_args
-        self.set_hessian(hessian)
-        self.set_structure(structure)
-        self.guess_windows(windows, window_frac)
-        self.set_noises(noises)
         self.M = M
         self.fit_kind = fit_kind
-        self.ls_list = self._generate_ls_list(mode = mode, pes_func = pes_func, **kwargs)
-        if mode == 'pes' and pes_func is not None and 'set_target' not in kwargs:  # FIXME: refactor set_target
-            self.loaded = True
-            self.calculate_next(**kwargs)
+        if structure is not None:
+            self.set_structure(structure)
+        #end if
+        if hessian is not None:
+            self.set_hessian(hessian)
+        #end if
+        if self.status.setup:
+            self.guess_windows(windows, window_frac)
+            self.set_noises(noises)
         #end if
     #end def
 
+    def _setup(self):
+        # test for proper hessian and structure
+        if isinstance(self.hessian, ParameterHessian) and isinstance(self.structure, ParameterSet):
+            # test for compatibility
+            if not len(self.hessian.Lambda) == len(self.structure.params):
+                return False
+            #end if
+        else:
+            return False
+        #end if
+        return PesSampler._setup(self)
+    #end def
+
+    # ie displacement of positions
+    def _shifted(self):
+        # check windows
+        if self.windows is None or not len(self.windows) == self.D or self.M is None:
+            return False
+        #end if 
+        # check noises
+        if self.noisy and self.noises is None:
+            return False
+        #end if
+        return PesSampler._shifted(self)
+    #end def
+
+    # def _generated(self): not overridden
+    def _loaded(self):
+        if not all([ls.loaded for ls in self.ls_list]):
+            return False
+        #end if
+        return PesSampler._loaded(self)
+    #end def
+
+    def _analyzed(self):
+        if not all([ls.analyzed for ls in self.ls_list]):
+            return False
+        #end if
+        return PesSampler._analyzed(self)
+    #end def
+
     def set_hessian(self, hessian):
-        self._protected()
-        if isinstance(hessian, ndarray):
+        self._avoid_protected()
+        if hessian is None:
+            return
+        elif isinstance(hessian, ndarray):
             hessian = ParameterHessian(hessian = hessian, x_unit = self.x_unit, E_unit = self.E_unit)
         elif isinstance(hessian, ParameterHessian):
             pass
@@ -92,15 +125,22 @@ class ParallelLineSearch():
         self.Lambdas = hessian.Lambda
         self.directions = hessian.get_directions()
         self.D = hessian.D
+        self.cascade()
     #end def
 
     def set_structure(self, structure):
-        self._protected()
+        self._avoid_protected()
+        if structure is None:
+            return
+        #end if 
         assert isinstance(structure, ParameterSet), 'Structure must be ParameterSet object'
         self.structure = structure.copy(label = 'eqm')
+        self.cascade()
     #end def
 
-    def guess_windows(self, windows, window_frac):
+    def guess_windows(self, windows, window_frac, **kwargs):
+        self._avoid_protected()
+        self._require_setup()
         if windows is None:
             windows = self.Lambdas**0.5 * window_frac
             self.windows_frac = window_frac
@@ -108,14 +148,20 @@ class ParallelLineSearch():
         self.set_windows(windows)
     #end def
 
-    def set_windows(self, windows):
+    def set_windows(self, windows, **kwargs):
+        self._avoid_protected()
+        self._require_setup()
         if windows is not None:
             assert windows is not None or len(windows) == self.D, 'length of windows differs from the number of directions'
             self.windows = array(windows)
         #end if
+        self.cascade()
+        self.reset_ls_list(**kwargs)  # always reset ls_list
     #end def
 
-    def set_noises(self, noises):
+    def set_noises(self, noises, **kwargs):
+        self._avoid_protected()
+        self._require_setup()
         if noises is None:
             self.noisy = False
             self.noises = None
@@ -124,12 +170,19 @@ class ParallelLineSearch():
             self.noisy = True
             self.noises = array(noises)
         #end if
+        self.cascade()
+        self.reset_ls_list(**kwargs)  # always reset ls_list
     #end def
 
-    def _generate_ls_list(self, **kwargs):
+    def reset_ls_list(self, D = None, **kwargs):
+        self._avoid_protected()
+        self._require_shifted()
+        if D is None:
+            D = range(self.D)
+        #end if
         noises = self.noises if self.noisy else self.D * [None]
         ls_list = []
-        for d, window, noise in zip(range(self.D), self.windows, noises):
+        for d, window, noise in zip(D, self.windows, noises):
             ls = LineSearch(
                 structure = self.structure,
                 hessian = self.hessian,
@@ -141,7 +194,14 @@ class ParallelLineSearch():
                 **kwargs)
             ls_list.append(ls)
         #end for
-        return ls_list
+        self.ls_list = ls_list
+        if self.mode == 'pes':
+            self.cascade()
+            self.load_results()
+        elif self.mode == 'files':
+            self.generate_jobs(**kwargs)
+        #end if
+        self.cascade()
     #end def
 
     def copy(self, path = '', c_noises = 1.0, **kwargs):
@@ -158,70 +218,131 @@ class ParallelLineSearch():
             'noises': noises,
             'M': self.M,
             'fit_kind': self.fit_kind,
-            'job_func': self.job_func,
-            'job_args': self.job_args,
-            'analyze_func': self.analyze_func,
+            'pes_func': self.pes_func,
+            'pes_args': self.pes_args,
+            'load_func': self.load_func,
+            'load_args': self.load_args,
+            'mode': self.mode,
         }
         ls_args.update(**kwargs)
-        ls_next = ParallelLineSearch(**ls_args)
-        return ls_next
+        pls_next = ParallelLineSearch(**ls_args)
+        return pls_next
     #end def
 
-    def propagate(self, path, protect = True, write = True, **kwargs):
-        assert self.calculated, 'Must calculate data before propagating linesearch'
-        pls_next = self.copy(path = path, structure = self.structure_next, **kwargs)
-        self.protected = protect
+    def propagate(self, path = None, protect = True, write = True, **kwargs):
+        if not self.status.analyzed:
+            return
+        #end if
+        path = path if path is not None else self.path + '_next/'
+        # check if manually providing structure
+        if 'structure' in kwargs.keys():
+            # TODO assert
+            pls_next = self.copy(path = path, **kwargs)
+        else:
+            pls_next = self.copy(path = path, structure = self.structure_next, **kwargs)
+        #end if
+        self.status.protected = protect
         if write:
             self.write_to_disk()
         #end if
         return pls_next
     #end def
 
-    def _protected(self):
-        assert not self.protected, 'ParallelLineSearch object is protected. Change at own risk!'
-    #end def
-
-    def generate_jobs(self, job_func = None, job_args = {}, **kwargs):
-        job_args = job_args if not job_args == {} else self.job_args
-        job_func = job_func if job_func is not None else self.job_func
+    def generate_jobs(self, pes_func = None, pes_args = {}, **kwargs):
+        self._require_shifted()
+        assert self.mode in ['nexus', 'files'], 'Supported only in nexus/files modes; present mode: {}'.format(self.mode)
+        pes_args = pes_args if not pes_args == {} else self.pes_args
+        pes_func = pes_func if pes_func is not None else self.pes_func
         sigma_min = None if not self.noisy else self.noises.min()
-        eqm_jobs = self.ls_list[0].generate_eqm_jobs(job_func, path = self.path, sigma = sigma_min, **job_args, **kwargs)
+        # TODO: check validity of pes_func
+        eqm_jobs = self.ls_list[0].generate_eqm_jobs(pes_func, path = self.path, sigma = sigma_min, **pes_args, **kwargs)
         jobs = eqm_jobs
         for ls in self.ls_list:
-            jobs += ls.generate_jobs(job_func, path = self.path, eqm_jobs = eqm_jobs, **job_args, **kwargs)
+             jobs += ls.generate_jobs(pes_func, path = self.path, eqm_jobs = eqm_jobs, **pes_args, **kwargs)
         #end for
-        self.generated = True
+        self.status.generated = True
         return jobs
     #end def
 
+    def run_jobs(self, interactive = True, **kwargs):
+        from nexus import run_project
+        jobs = self.generate_jobs(**kwargs)
+        if jobs is None or jobs == []:
+            return
+        #end if
+        if interactive:
+            print('About to run:')
+            for job in jobs:
+                #print(job)
+                print('  {}'.format(job.path))
+            #end for
+            if input('proceed? (Y/n) ') in ['n', 'N']:
+                exit()
+            else:
+                run_project(jobs)    
+            #end if
+        else:
+            run_project(jobs)    
+        #end if
+    #end def
+
     # can either load based on analyze_func or by providing values/errors
-    def load_results(self, analyze_func = None, values = None, errors = None, **kwargs):
-        self._protected()
-        analyze_func = analyze_func if analyze_func is not None else self.analyze_func
-        values_ls = values if values is not None else self.D * [None]
-        errors_ls = errors if errors is not None else self.D * [None]
+    def load_results(self, load_func = None, values = None, errors = None, **kwargs):  # load_args
+        if self.status.protected:
+            return
+        #end if
         loaded = True
+        if self.mode == 'pes' and self.pes_func is not None:
+            load_func = None
+            values_ls, errors_ls = [], []
+            for ls in self.ls_list:
+                res = ls.evaluate_pes(pes_func = self.pes_func)
+                values_ls.append(res[1])
+                errors_ls.append(res[2])
+            #end for
+        else:
+            load_func = load_func if load_func is not None else self.load_func
+            values_ls = values if values is not None else self.D * [None]
+            errors_ls = errors if errors is not None else self.D * [None]
+        #end if
         for ls, values, errors in zip(self.ls_list, values_ls, errors_ls):
             loaded_this = ls.load_results(
-                analyze_func = analyze_func,
+                load_func = load_func,
                 values = values,
                 errors = errors,
                 path = self.path,
                 **kwargs)
             loaded = loaded and loaded_this
         #end for
-        self.loaded = loaded
         if loaded:
+            self.find_eqm_value()
+            self.status.generated = True
+            self.status.loaded = True
             self.calculate_next()
         #end if
+        self.cascade()
+    #end def
+
+    def find_eqm_value(self):
+        E, Err = None, None
+        for ls in self.ls_list:
+            for s in ls.structure_list:
+                if sum((self.structure.params - s.params)**2) < 1e-10:
+                    E, err = s.value, s.value_err
+                    break
+                #end if
+            #end for
+        #end for
+        self.structure.value = E
+        self.structure.value_err = err
     #end def
 
     def calculate_next(self, **kwargs):
-        self._protected()
-        assert self.loaded, 'Must load data before calculating next parameters'
+        self._avoid_protected()
+        self._require_loaded()
         params_next, params_next_err = self.calculate_next_params(**kwargs)
         self.structure_next = self.structure.copy(params = params_next, params_err = params_next_err)
-        self.calculated = True
+        self.cascade()
     #end def
 
     def ls(self, i):
@@ -276,7 +397,7 @@ class ParallelLineSearch():
     #end def
 
     def get_shifts(self):
-        assert self.loaded, 'Shift data has not been calculated yet'
+        self._require_loaded()
         return self._get_shifts()
     #end def
 
@@ -306,7 +427,11 @@ class ParallelLineSearch():
         return array([ls.get_x0_distribution(errors = ls.errors, N = N, **kwargs) for ls in self.ls_list]).T
     #end def
 
-    def write_to_disk(self, fname = 'data.p'):
+    def write_to_disk(self, fname = 'data.p', overwrite = False):
+        if path.exists(self.path + fname) and not overwrite:
+            #print('File {} exists. To overwrite, run with overwrite = True'.format(self.path + fname))
+            return
+        #end if
         makedirs(self.path, exist_ok = True)
         with open(self.path + fname, mode='wb') as f:
             f.write(dumps(self))
